03_generation_with_context/requirements.txt
langchain>=0.3.24
langchain-openai>=0.1.6
langchain-qdrant>=0.2.0
openai>=1.0.0
qdrant-client>=1.14.0
python-dotenv>=1.0.1
